Heroku deployment
There were several changes required to get the Heroku deployment in a
working state:
* Move api-server/server.js to the root directory
* Move api-server/package.json to the root directory

These two changes were needed to provide a sane entry point for Heroku.
As part of this change, I also did quite a bit of renaming:
* moved api-server/config.js to the root
* `api-server` to `api`
* `frontend` to `client`

Finally, I used the `proxy` directive in `client/package.json` to have
the client work with the API a bit more clearly.